force reval of target when following LAST_BIND symlinks
Patchwork-id: 22023
O-Subject: [RHEL6.0 PATCH] BZ#548153: vfs: force reval of target when following
	LAST_BIND symlinks
Bugzilla: 548153
RH-Acked-by: Peter Staubach <staubach@redhat.com>

This patch just went to mainline and will be in 2.6.33.

procfs-style symlinks return a last_type of LAST_BIND without an actual
path string. This causes __follow_link to skip calling __vfs_follow_link
and so the dentry isn't revalidated.

This is a problem when the link target sits on NFSv4 as it depends on
the VFS to revalidate the dentry before using it on an open call. Ensure
that this occurs by forcing a revalidation of the target dentry of
LAST_BIND symlinks.

Signed-off-by: Jeff Layton <jlayton@redhat.com>
Acked-by: "Eric W. Biederman" <ebiederm@xmission.com>
Acked-by: Miklos Szeredi <mszeredi@suse.cz>
Signed-off-by: Al Viro <viro@zeniv.linux.org.uk>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>