cciss: remove pci-ids
Patchwork-id: 22085
O-Subject: [RHEL6 2/2] bz464649 cciss: remove pci-ids
Bugzilla: 464649
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

This is for bz#464649

This patch comes from HP. hpsa is a new
driver which should be used in the future instead
of cciss for the newer(not older than 4 years)
HP controllers. The older cciss will still be
used for the older cards.

Almost the whole driver is now in scsi-misc tree,
there are some minor differences from last days.
The chances are good it will made it into 2.6.33
It is important for us to use this driver in RHEL6,
because of the long term upstream support.

I've tested this only slightly on an unsupported
older controller (it worked) and on one newer one.
We hope the main testing will come from HP.

First patch contains the new driver - hpsa,
the second one removes pci-id's from the cciss driver.
There is on controller left behind - it's the P400i
we don't know yet which driver should handle that board.
I'm working with HP on that and will come later with
a patch.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>