support MAXSMP
Patchwork-id: 22107
O-Subject: [RHEL6 PATCH] [xen] [pv-ops domU] support MAXSMP
Bugzilla: 547129
RH-Acked-by: Paolo Bonzini <pbonzini@redhat.com>
RH-Acked-by: Miroslav Rezanina <mrezanin@redhat.com>

Resolves BZ 547129.

The MAXSMP config option requires CPUMASK_OFFSTACK, which in turn
requires we init the memory for the maps while we bringing up the cpus.
MAXSMP also increases NR_CPUS to 4096. This increase in size exposed an
issue in the argument construction for mulitcalls from
xen_flush_tlb_others. The args should only need space for the actual
number of cpus, which with xen is currently only up to 32.

This patch has been submitted upstream and should be included in
Jeremy's tree soon.

The patch allows the RHEL6 kernel to boot as a PV guest. Tested by me.
I had problems with brew building, so I don't have a brew build.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>