utrace core
Patchwork-id: 22119
O-Subject: [RHEL6.0 PATCH 16/16] utrace core
Bugzilla: 549491
RH-Acked-by: Oleg Nesterov <oleg@redhat.com>

This adds the utrace facility, a new modular interface in the kernel
for implementing user thread tracing and debugging.  This fits on top
of the tracehook_* layer, so the new code is well-isolated.

The new interface is in <linux/utrace.h> and the DocBook utrace book
describes it.  It allows for multiple separate tracing engines to work
in parallel without interfering with each other.  Higher-level tracing
facilities can be implemented as loadable kernel modules using this layer.

The new facility is made optional under CONFIG_UTRACE. It can only be
enabled on machines that have all the prerequisites and select
CONFIG_HAVE_ARCH_TRACEHOOK.

If CONFIG_UTRACE is set ptrace uses the utrace facilities, it can play
nicely with other utrace-based things tracing the same threads.

Signed-off-by: Oleg Nesterov <oleg@redhat.com>
Signed-off-by: Roland McGrath <roland@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>