cio: device recovery stalls after multiple hardware events
Patchwork-id: 22128
O-Subject: [RHEL6 PATCH 02/13] [s390] cio: device recovery stalls after multiple
	hardware events
Bugzilla: 548490
RH-Acked-by: Pete Zaitcev <zaitcev@redhat.com>

Description
-----------
When an I/O device is detached and later re-attached, it could
sometimes remain detached in Linux.

Hardware events are handled on the subchannel layer, while attachment
status is handled on the device layer.  This design can lead to
situations where multiple attachments are scheduled for the same
subchannel, resulting in one or more inaccessible devices.

The solution is to handle the attachment status on the subchannel layer.

Bugzilla
--------
BZ 548490
https://bugzilla.redhat.com/show_bug.cgi?id=548490

Upstream status of the patch
----------------------------
The patch is upstream as of kernel version 2.6.33
http://git.kernel.org/?p=linux/kernel/git/torvalds/linux-2.6.git;a=commitdiff;h=5d6e6b6f6f3eac10a7f5a15e961bac3b36824d9d
http://git.kernel.org/?p=linux/kernel/git/torvalds/linux-2.6.git;a=commitdiff;h=60e4dac1abdf49ccdb7545ec406325f08423d848

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.

Please ACK.

With best regards,

	Hendrik

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>