cio: kernel panic after unexpected interrupt
Patchwork-id: 22130
O-Subject: [RHEL6 PATCH 06/13] [s390] cio: kernel panic after unexpected
	interrupt
Bugzilla: 548490
RH-Acked-by: Pete Zaitcev <zaitcev@redhat.com>

Description
-----------
Unexpected kernel panic with message
  "kernel BUG at drivers/s390/cio/device_fsm.c"
occurs after an I/O device entered not operational state.

When a device becomes temporarily not operational in a way that it
cannot be cleanly shut down, it could later report interrupts which
are unexpected and cause a kernel panic.

Repeat the shut down procedure when unexpected interrupts are
received for devices which are not operational.

Bugzilla
--------
BZ 548490
https://bugzilla.redhat.com/show_bug.cgi?id=548490

Upstream status of the patch
----------------------------
The patch is upstream as of kernel version 2.6.33
http://git.kernel.org/?p=linux/kernel/git/torvalds/linux-2.6.git;a=commitdiff;h=16b9a0571da4ee5cd15ca75e871722b0b5aee64d

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.

Please ACK.

With best regards,

	Hendrik

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>