cio: not operational devices cannot be deactivated
Patchwork-id: 22137
O-Subject: [RHEL6 PATCH 08/13] [s390] cio: not operational devices cannot be
	deactivated
Bugzilla: 548490
RH-Acked-by: Pete Zaitcev <zaitcev@redhat.com>

Description
-----------
Attempts to set a device offline which has become not operational
are rejected with message "Resource temporarily unavailable".

A check in the device offline function rejects the offline
request in case the device is not in operational state.

Remove the check for not operational so that users can deactivate
and remove devices from Linux which, at some point, appeared
not operational.

Bugzilla
--------
BZ 548490
https://bugzilla.redhat.com/show_bug.cgi?id=548490

Upstream status of the patch
----------------------------
The patch is upstream as of kernel version 2.6.33
http://git.kernel.org/?p=linux/kernel/git/torvalds/linux-2.6.git;a=commitdiff;h=350e91207bc9c6a464c22b9e0e30d21dfc07efe3

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.

Please ACK.

With best regards,

	Hendrik

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>