cio: memory leaks when checking unusable devices
Patchwork-id: 22138
O-Subject: [RHEL6 PATCH 12/13] [s390] cio: memory leaks when checking unusable
	devices
Bugzilla: 548490
RH-Acked-by: Pete Zaitcev <zaitcev@redhat.com>

Description
-----------
The amount of free memory is reduced with each time an I/O device
is checked that is not operational.  A reference counting problem
occurs when a device does not pass device recognition.

This patch corrects the reference counting.

Bugzilla
--------
BZ 548490
https://bugzilla.redhat.com/show_bug.cgi?id=548490

Upstream status of the patch
----------------------------
The patch is upstream as of kernel version 2.6.33
http://git.kernel.org/?p=linux/kernel/git/torvalds/linux-2.6.git;a=commitdiff;h=7d253b9a1aaf5192808e641659f4feb122faa536
http://git.kernel.org/?p=linux/kernel/git/torvalds/linux-2.6.git;a=commitdiff;h=24a1872d6411c7cce82c0888a4fbea23e993e051

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.

Please ACK.

With best regards,

	Hendrik

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>