sync fcoe with upstream
Patchwork-id: 22158
O-Subject: [PATCH RHEL 6.0]: sync fcoe with upstream
Bugzilla: 549945
RH-Acked-by: Rob Evers <revers@redhat.com>
RH-Acked-by: Tomas Henzl <thenzl@redhat.com>

From: Mike Christie <mchristi@redhat.com>

This is for BZ 462791.

This updates the fcoe tree to what is in the scsi maintainer's
scsi-miscs tree
http://git.kernel.org/?p=linux/kernel/git/jejb/scsi-misc-2.6.git;a=summary

This patch depends on Andy's net/ixbe update. I had done
some testing with a earlier version of the net/ixgbe update,
and found some problems:
1. Performance regression. Intel debugged this and it was
due to a config/setup problem.

2. When I insmod ixgbe ethX does not always show up. This
was not a regression caused by the patches. Intel is working
on it and we will send a patch for that in a new bz when
it is resolved.

Now, I do not have any setup. I am in the middle of moving,
again, so I am trusting Intel's testing.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>