ixgbe: add support for 82599-KR and update to latest upstream
Patchwork-id: 22165
O-Subject: [RHEL6.0 PATCH] ixgbe: add support for 82599-KR and update to
	latest upstream
Bugzilla: 462781
RH-Acked-by: David S. Miller <davem@redhat.com>

This patch updates the ixgbe driver to the latest in Linus' tree.  All
commits through this one were included:

commit ca553980432898da5d4125573a9e2aee6ed5d355
Author: Gurucharan Shetty <gshetty@riverbed.com>
Date:   Tue Dec 15 13:00:31 2009 +0000

    ixgbe: allow tx of pre-formatted vlan tagged packets

I have not tested this on a RHEL6 kernel, but this code is identical to
the latest upstream (and have run that) I feel confident posting it.

This will resolve RHBZ 462781.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>