scsi_dh: Make rdac hardware handler s activate async
Patchwork-id: 22189
O-Subject: [rhel6.0 patch 2/4 v2] scsi_dh:
	Make-rdac-hardware-handler-s-activate-async
Bugzilla: 537257
RH-Acked-by: Mike Christie <mchristi@redhat.com>

https://bugzilla.redhat.com/show_bug.cgi?id=537257

Batch up MODE_SELECT in rdac device handler.

LSI RDAC storage has the capability of handling mode selects for
multiple luns in a same command. Make use of that ability to send
as few MODE SELECTs as possible to the storage controller as possible.

This patch creates a work queue and queues up activate requests
when a MODE SELECT is sent down the wire. When that MODE SELECT
completes, it compiles queued up activate requests for multiple
luns into a single MODE SELECT.

This reduces the time to do failover/failback of large number of LUNS.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>