add feature flags to reduce future kABI impact
Patchwork-id: 22202
O-Subject: [RHEL6 PATCH] dm: add feature flags to reduce future kABI impact
Bugzilla: 547756
RH-Acked-by: Jonathan E Brassow <jbrassow@redhat.com>

BZ 547756

This patch adds a 'feature flags' field to several structures that we
anticipate we may wish to extend for enhancements during the lifetime of the
product, to help us reduce the impact of future kABI changes.
(Mostly these are lists of function pointers that we might need to
extend if new features require new hooks)

This patch deals with a problem peculiar to RHEL and is not currently
upstream.

Signed-off-by: Mike Snitzer <snitzer@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>