make driver PCI legacy I/O port free
Patchwork-id: 22242
O-Subject: [RHEL6 PATCH] bz#549118 megaraid_sas: make driver PCI legacy I/O
	port free
Bugzilla: 549118
RH-Acked-by: David Milburn <dmilburn@redhat.com>

This is for bz#549118

The patch is in 2.6.33.rc1
http://git.kernel.org/?p=linux/kernel/git/torvalds/linux-2.6.git;a=commit;h=aeab3fd7b865bc4086a80a83cfdd67dded3b41a0

We have added this to RHEL5.5, so it would be
good to have it also in RHEL6.

I've only compile tested, Fujitsu has promised to
test the RHEL5.5 version, I expect they will also test
this one.

Tomas

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>