ext4: flush delalloc blocks when space is low
Patchwork-id: 22279
O-Subject: [PATCH RHEL6 2/2] ext4: flush delalloc blocks when space is low
Bugzilla: 526758
RH-Acked-by: Peter Staubach <staubach@redhat.com>

This is for Bug 526758 - ext4: Better ENOSPC handling needed

This patch has been sent upstream, acked by Jens Axboe and
Jan Kara, and merged into the ext4 patch queue but is not
yet technically upstream.

===

Creating many small files in rapid succession on a small
filesystem can lead to spurious ENOSPC; on a 104MB filesystem:

for i in `seq 1 22500`; do
    echo -n > $SCRATCH_MNT/$i
    echo XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX > $SCRATCH_MNT/$i
done

leads to ENOSPC even though after a sync, 40% of the fs is free
again.

This is because we reserve worst-case metadata for delalloc writes,
and when data is allocated that worst-case reservation is not
usually needed.

When freespace is low, kicking off an async writeback will start
converting that worst-case space usage into something more realistic,
almost always freeing up space to continue.

This resolves the testcase for me, and survives all 4 generic
ENOSPC tests in xfstests.

We'll still need a hard synchronous sync to squeeze out the last bit,
but this fixes things up to a large degree.

Signed-off-by: Eric Sandeen <sandeen@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>