cciss: fix spinlock use
Patchwork-id: 22324
O-Subject: [RHEL6 PATCH] bz#552910 cciss: fix spinlock use
Bugzilla: 552910
RH-Acked-by: Rob Evers <revers@redhat.com>
RH-Acked-by: Pete Zaitcev <zaitcev@redhat.com>

This is for bz#552910

fail_all_cmds is called (only) from do_cciss_intr with
spinlock acquired.
So when in an error-handling situation,
fail_all_cmds() is called it attempts to *reacquire* the lock.

This patch removes the spinlock lock/unlock from fail_all_cmds and adds a
spin_unlock after the call to fail_all_cmds before return.

This patch is in RHEl5, it was posted posted in linux-scsi and ack-ed there
by Mike Miller (HP), but is not yet in any upstream tree.

Tomas

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>