GFS2: Fix up system xattrs
Patchwork-id: 22365
O-Subject: [PATCH 01/08] GFS2: Fix up system xattrs
Bugzilla: 546294
RH-Acked-by: Robert S Peterson <rpeterso@redhat.com>

This code has been shamelessly stolen from XFS at the suggestion
of Christoph Hellwig. I've not added support for cached ACLs so
far... watch for that in a later patch, although this is designed
in such a way that they should be easy to add.

Signed-off-by: Steven Whitehouse <swhiteho@redhat.com>
Cc: Christoph Hellwig <hch@infradead.org>