GFS2: Add cached ACLs support
Patchwork-id: 22368
O-Subject: [PATCH 06/08] GFS2: Add cached ACLs support
Bugzilla: 546294
RH-Acked-by: Robert S Peterson <rpeterso@redhat.com>

The other patches in this series have been building towards
being able to support cached ACLs like other filesystems. The
only real difference with GFS2 is that we have to invalidate
the cache when we drop a glock, but that is dealt with in earlier
patches.

Signed-off-by: Steven Whitehouse <swhiteho@redhat.com>