tape: Add pr_fmt() macro to all tape source files
Patchwork-id: 22427
O-Subject: [RHEL6 PATCH 1/1] [s390] tape: Add pr_fmt() macro to all tape
	source files
Bugzilla: 554380

Description
-----------
The "tape: " prefix is missing in the output of pr_xxx() printk
macros.  This is caused by a missing pr_fmt() macro definition.

The solution is to add the definition of the pr_fmt() containing
the prefix.

Bugzilla
--------
BZ 554380
https://bugzilla.redhat.com/show_bug.cgi?id=554380

Upstream status of the patch
----------------------------
The patch is upstream as of kernel version 2.6.33
http://git.kernel.org/?p=linux/kernel/git/torvalds/linux-2.6.git;a=commitdiff;h=bb509912481214cf6ad1181c968295c62ff1ad9e

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.