SVM: Move INTR vmexit out of atomic code
Patchwork-id: 22470
O-Subject: [RHEL 6 KVM 08/20] KVM: SVM: Move INTR vmexit out of atomic code
Bugzilla: 554762
RH-Acked-by: Avi Kivity <avi@redhat.com>
RH-Acked-by: Gleb Natapov <gleb@redhat.com>
RH-Acked-by: Juan Quintela <quintela@redhat.com>

commit cd3ff653ae0b45bac7a19208e9c75034fcacc85f
Author: Joerg Roedel <joerg.roedel@amd.com>
Date:   Fri Oct 9 16:08:26 2009 +0200

KVM: SVM: Move INTR vmexit out of atomic code

The nested SVM code emulates a #vmexit caused by a request
to open the irq window right in the request function. This
is a bug because the request function runs with preemption
and interrupts disabled but the #vmexit emulation might
sleep. This can cause a schedule()-while-atomic bug and is
fixed with this patch.

Signed-off-by: Joerg Roedel <joerg.roedel@amd.com>
Signed-off-by: Marcelo Tosatti <mtosatti@redhat.com>
Bugzilla: 554762

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>