SVM: Notify nested hypervisor of lost event injections
Patchwork-id: 22475
O-Subject: [RHEL 6 KVM 09/20] KVM: SVM: Notify nested hypervisor of lost event
	injections
Bugzilla: 554762
RH-Acked-by: Avi Kivity <avi@redhat.com>
RH-Acked-by: Gleb Natapov <gleb@redhat.com>
RH-Acked-by: Juan Quintela <quintela@redhat.com>

commit 8d23c4662427507f432c96ac4fa3b76f0a8360cd
Author: Alexander Graf <agraf@suse.de>
Date:   Fri Oct 9 16:08:25 2009 +0200

KVM: SVM: Notify nested hypervisor of lost event injections

If event_inj is valid on a #vmexit the host CPU would write
the contents to exit_int_info, so the hypervisor knows that
the event wasn't injected.

We don't do this in nested SVM by now which is a bug and
fixed by this patch.

Signed-off-by: Alexander Graf <agraf@suse.de>
Signed-off-by: Joerg Roedel <joerg.roedel@amd.com>
Signed-off-by: Marcelo Tosatti <mtosatti@redhat.com>
Bugzilla: 554762

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>