Allow internal errors reported to userspace to carry extra data
Patchwork-id: 22472
O-Subject: [RHEL 6 KVM 15/20] KVM: Allow internal errors reported to userspace
	to carry extra data
Bugzilla: 554762
RH-Acked-by: Avi Kivity <avi@redhat.com>
RH-Acked-by: Gleb Natapov <gleb@redhat.com>
RH-Acked-by: Juan Quintela <quintela@redhat.com>

commit a9c7399d6cda0a092b347f8ee49bbe44f6e1fe66
Author: Avi Kivity <avi@redhat.com>
Date:   Wed Nov 4 11:54:59 2009 +0200

KVM: Allow internal errors reported to userspace to carry extra data

Usually userspace will freeze the guest so we can inspect it, but some
internal state is not available.  Add extra data to internal error
reporting so we can expose it to the debugger.  Extra data is specific
to the suberror.

Signed-off-by: Avi Kivity <avi@redhat.com>
Bugzilla: 554762

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>